Adding to Duarte Loreto as Portuguese translator.
Fatima has told me there is not need to add the Portuguese (Brazil) translation to the site.  She says that as there are not two English languages, there are not two Portuguese languages. She is Brazilian, and she says Brazilian people will be able to read more or less the site text.