Fixed the below two validator.w3.org errors:
# Error  Line 268 column 507: end tag for element "P" which is not open.

...Free Software</a> philosophy.</ul></p>

The Validator found an end tag for the above element, but that element is not currently open. This is often caused by a leftover end tag from an element that was removed during editing, or by an implicitly closed element (if you have an error related to an element being used where it is not allowed, this is almost certainly the case). In the latter case this error will disappear as soon as you fix the original problem.

If this error occured in a script section of your document, you should probably read this FAQ entry.



# Error Line 292 column 233: end tag for element "P" which is not open.

... non-technical subjects.</li></ul></p