We use HTTPS for all webpages.
As exposed by Victor, it is not needed add "/".  Note that at the Apache virtual domain we add "/" if it is needed:
   Redirect / https://www.gnuherds.org/

Note: It is not needed because all web pages are kept in the same directory, not in subdirectories.  So, we do not have to add "/" to come back from subdirectories to the site entry point directory.