Set alerts on NewJobOffers defaults to 'true'
We change the default value of Alerts-on-NewJobOffers from 'false' to 'true' due to some entities do not manage to find it and set it to 'true'.

If for example a company register, and they do not want to get this type of alerts, they can turn off it easily after receiving the first alert, due to the own email alert includes the instructions to turn of it.