Offers look as a resumes instead of job offer! Fixing it.
Some guys exposed that the offer view looked as a resume instead of an offer!

If you read it carefully you note it's a job offer. But at first it looks like a resume, for sure.

The fact that it starts out with a person's name and when he was born, and "is looking for vacancy" and then has phrases like "experience since" and other sorts of details that one normally sees in a resume, that's what confuses.

That should be changed. If it started out "wanted: programmer" then it'd be less ambiguous.

Following the advice of such guys, I have moved the 'employer' section at the end.