The licenses allow us to apply the below license upgrade path even without being the copyright holder, due to the  "or (at your option) any later version" text is present at each licensed file:
  AGPLv1   -->   AGPL2   -->   GNU AGPLv3

* AGPLv1,     http://www.affero.org/oagpl.html
* AGPL2,      http://www.affero.org/agpl2.html
* GNU AGPLv3, http://www.fsf.org/licensing/licenses/agpl-3.0.html

Any change that is not allowed by the "license upgrade path" must be
explicitly agreed by _all_ copyright holders.

We move directly from AGPLv1 to GNU AGPLv3 and just expose here that we have used AGPL2 as the upgrade bridge.
  Reference to the development email list: http://lists.gnu.org/archive/html/gnuherds-app-dev/2007-11/msg00056.html

AGPL2,      http://www.affero.org/agpl2.html
=====

Note: The GNU Affero General Public License version 3 has been released. It follows the same basic structure as version 1 of the Affero GPL, but it is based on version 3 of the GNU GPL, and you can combine modules released under GPLv3 and the GNU AGPLv3. If you were thinking about releasing your software under this license, we recommend you use the GNU AGPLv3 instead.

AFFERO GENERAL PUBLIC LICENSE

Version 2, November 2007

Copyright © 2007 Affero Inc.
510 Third Street - Suite 225, San Francisco, CA 94107, USA

This is version 2 of the Affero General Public License. It gives each licensee permission to distribute the Program or a work based on the Program (as defined in version 1 of the Affero GPL) under the GNU Affero General Public License, version 3 or any later version.

If the Program was licensed under version 1 of the Affero GPL "or any later version", no additional obligations are imposed on any author or copyright holder of the Program as a result of a licensee's choice to follow this version 2 of the Affero GPL.