IE 6.0 and IE 7.0 <button> bug workaround
IE does not work with <button>s as the specification expose [1]
  [1] http://www.w3.org/TR/REC-html40/interact/forms.html#edef-BUTTON

It's supposed to submit the value attribute, as any other browser does following the specification, but it instead submits the contents of the element!

It is even worse, IE sends _always_ the content of _all_ <button>s, even if buttons have not been clicked!

Work around: Use <input> instead of <button> when a IE browser is detected.