Do not fill a password if the user do not fill it
To avoid confuse the user the password must not be filled up to the user fill it. Note the user can not read the value it keeps.  The value remembered by FireFox is a different case due to the user ask for such 'remembering' and [s]he is confident about its correctness.