Avoid FireFox showing the password value in the password field of Person, Company and non-profit edition forms when the user ask for remembering the log-in email and password
To avoid it we add a prefix to the log-in box fields, so they do not have the same name than the one of Person, Company and non-profit edition forms.

Note: We want avoid it due to when the password value is shown in the edition form it makes it harder to the user to save the form because of only one of the password field will keep the value and so the checks-before-saving will fail with the message: "Password fields differ".