Set the NewJobOffer alert flag to 'true' by default
So entities, that is to say, persons, companies and non-profits will receive the NewJobOffer alert email up to they change such option in the Settings menu.
Note: The alerts emails explain how to disable it.

That is a good thing, due to new entities (users of the webapp) will not know about such option, so they could not enable it.