Bug fix: Bad usability at filling dates
Error messages read: "Incorrect date format" but it does not say what the correct format is.
Thanks to MJ Ray for reporting.

Usability: Besides, we should add more regular expression to allow fill '1994-9-1' and not force to fill '1994-09-01'.

Maybe we could improve the way to fill the dates, due to the current one is maybe not the best one.