Spanish translation: partial update
I just wanted to do a partial update of the Spanish translation to know
how it looked, and so at the same time check the English text.  Note
that as usual the translation will not be carried out up to we finish
all the pendind development for this development phase.