Donation-pledge-groups: Fix the manage-my-donations feature
My donations are only listed if I was who created the
DonationPledgeGroup.

This commit is not a bug fix. This commit replace some code which
was just wrong.  That is to say, for donations we must list all the
donations and then complete with the DonationPledgeGroup title, etc.,
but not list the DonationPledgeGroup created by the logged entity and
then complete with his/her donations.