Improve usability: Do not kill the user log-in session
If the webapp send an email with an HTTP://gnuherds.org/ link, instead of
a HTTPS one, when the user click it the webapp will open a new PHP session
killing the previous one if the user was already logged.  So, the webapp
is forcing him to log in again.

If the webapp send to the user an HTTPS link a new PHP session is not
created, so the user can follow using his previous PHP session without
being forced to log in again.