Bug fix: break adding new donation-pledge-group or look-for-volunteers entries
The webapp breaks adding a new donation-pledge-group or look-for-volunteers entry with
the below message due to addEntity() needs the magic number but it has not been supplied.

This only happen if the operation is done by a non-registered email. When the webapp
tries to register the email it fails due to it miss the 'magic' number.

  ERROR:

     PREPARE query(text,integer) AS  UPDATE E1_Entities SET E1_RegisterMagic=$1, E1_RegisterMagicExpire= now() + '2 days'::interval WHERE E1_Id=$2;  EXECUTE query('',2);

  ERROR: new row for relation "e1_entities" violates check constraint "e1_entities_e1_registermagic_check"